Output of `npm run bench` with these changes:

```
parse .../node_modules/@scope/name/index.js   x 9,263,589 ops/sec (10 runs sampled) v8-never-optimize=true min..max=(104.84ns...112.00ns)
parse /path/to/node_modules/name/index.js     x 12,193,545 ops/sec (9 runs sampled) v8-never-optimize=true min..max=(78.95ns...84.14ns)
parse /path/to/node_modules/name/sub/index.js x 10,707,062 ops/sec (10 runs sampled) v8-never-optimize=true min..max=(90.45ns...95.77ns)
parse .../invalid/node_modules/name/index.js  x 10,529,465 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(93.12ns...97.64ns)
parse /path/to/node_modules                   x 33,774,271 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(28.41ns...30.32ns)
parse /path/to/no/module                      x 29,484,475 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(33.41ns...34.98ns)
parse                                         x 34,020,946 ops/sec (10 runs sampled) v8-never-optimize=true min..max=(28.77ns...30.98ns)
```

Without these changes:

```
parse .../node_modules/@scope/name/index.js   x 5,366,664 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(184.70ns...189.33ns)
parse /path/to/node_modules/name/index.js     x 6,200,375 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(156.56ns...172.60ns)
parse /path/to/node_modules/name/sub/index.js x 5,616,072 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(172.03ns...189.01ns)
parse .../invalid/node_modules/name/index.js  x 5,604,977 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(174.59ns...184.59ns)
parse /path/to/node_modules                   x 34,420,842 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(28.89ns...29.21ns)
parse /path/to/no/module                      x 29,450,337 ops/sec (11 runs sampled) v8-never-optimize=true min..max=(33.41ns...34.91ns)
parse                                         x 33,930,821 ops/sec (10 runs sampled) v8-never-optimize=true min..max=(28.82ns...29.95ns)
```